+++ searcher/policy.go
@@ -13,17 +13,17 @@ type uct struct {
 	numerator float64
 }
 
-func newUCT(cSquared float64, N int) *uct {
+func newUCT(cSquared float64, N float64) *uct {
 	if N == 0 {
 		panic("N cannot be 0")
 	}
-	return &uct{numerator: cSquared * math.Log(float64(N))}
+	return &uct{numerator: cSquared * math.Log(N)}
 }
 
-func (u uct) evaluate(q float64, n int) float64 {
+func (u uct) evaluate(q float64, n float64) float64 {
 	if n == 0 {
 		panic("n cannot be 0")
 	}
 	// UCT = q/n + sqrt(c^2*ln(N)/n)
-	return q/float64(n) + math.Sqrt(u.numerator/float64(n))
+	return q/n + math.Sqrt(u.numerator/n)
 }
